+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Fix duplicates ending up in `Toot.reblogsBy` on page reloads
 * Add `DISCUSSIONS` preset; tweak `TRENDING` preset
 
 ### v0.17.3